cheat-sheet: Show copy-success message momentarily
[why]
It can be confusing to click on the pseudo-button (to copy the codepoint
or whatever) and it does not show anything happened at all.

[how]
Change the pseudo-button title to 'ok' momentarily (i.e. for 1 second).

Also correct message when copy-to-clipboard is disabled and no exception
is thrown but false is returned.

Fixes: #1764

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>