cheatsheet: Do not remove dashes from index
[why]
When people want to search for one complete set, i.e. "nf-pl-" this is
not really possible (or gives many wrong positives) because the dashes
are used as splitting points; you end up searching for "nf" and "pl".

[how]
We have no whitespace in any text we index, and we have '-' as set
delimiter and '_' as blank/whitespace simulator.

So we split now the keys and the search pattern on blank and underscore.
This allows correct search for "nf-pl-" which was impossible before.

Fixes: #1924

Reported-by: @Immortal-Sty
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>