Fix showing all icons
[why]
The search term ' ' to show all icons does not work anymore after the
last commit
  71dfe1296  cheatsheet: Do not remove dashes from index

The reason is that before we split the names into terms, like
`nf-pl-left_hard_divider` has been split into `nf`, `pl`, `left`,
`hard`, and `divider`.

But that made it impossibe to search for (only) `pl-` to show all
powerline icons.

That commit changed the tokenization such that we get instead
`nf-pl-left`, `hard`, and `divider`.

Before we searched to find all icons for `nf` as exact hit, has that was
always part of any icon. But that is not the case anymore.
Now we need to search for `nf-` and that as prefix hit.

[note]
From the above it seems impossible to search for `left`, as that is part
of the new first token, but at its end.

But that is not a problem because we do not only index the tag, but also
the name, which is `left_hard_divider` -> `left`, `hard`, `divider`.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>